♻️ refactor: 코드 정리 및 주석추가
