+++ src/main/java/com/example/newsfeedproject/users/dto/SearchUserResponseDto.java
@@ -1,8 +1,10 @@
 package com.example.newsfeedproject.users.dto;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
+@NoArgsConstructor
 public class SearchUserResponseDto {
     private String userName;
     private String profileImageUrl;
@@ -13,5 +15,4 @@ public class SearchUserResponseDto {
         this.profileImageUrl = profileImageUrl;
         this.followed = followed;
     }
-
 }
